Policyを追加

+++ app/Http/Controllers/TodoController.php
@@ -7,26 +7,17 @@ use Illuminate\Http\Request;
 
 class TodoController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         $todos = auth()->user()->todos()->latest()->get();
         return view('todos.index', compact('todos'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
         return view('todos.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         //バリデーション
@@ -41,27 +32,21 @@ class TodoController extends Controller
         return redirect()->route('todos.index')->with('success', 'ToDoを作成しました。');
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Todo $todo)
     {
         abort(404);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(Todo $todo)
     {
+        $this->authorize('update', $todo);
         return view('todos.edit', compact('todo'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Todo $todo)
     {
+        $this->authorize('update', $todo);
+
         //バリデーション
         $validated = $request->validate([
             'title' => 'required|string|max:255',
@@ -72,11 +57,9 @@ class TodoController extends Controller
         return redirect()->route('todos.index')->with('success', '更新しました');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Todo $todo)
     {
+        $this->authorize('delete', $todo);
         $todo->delete();
 
         return redirect()->route('todos.index')->with('success', 'ToDoを削除しました。');
